Add tests for lock plugin
