+++ conda-store-server/tests/_internal/action/test_actions.py
@@ -11,7 +11,6 @@ import tempfile
 from unittest import mock
 
 import pytest
-import yaml
 import yarl
 from celery.result import AsyncResult
 from conda.base.context import context as conda_base_context
@@ -23,7 +22,6 @@ from conda_store_server import BuildKey, api
 from conda_store_server._internal import action, conda_utils, orm, schema, server, utils
 from conda_store_server._internal.action import (
     generate_constructor_installer,
-    generate_lockfile,
 )
 from conda_store_server.server.auth import DummyAuthentication
 
@@ -70,114 +68,6 @@ def test_action_decorator():
     assert context.result != pathlib.Path.cwd()
     # test that temporary directory is cleaned up
     assert not context.result.exists()
-
-
-@pytest.mark.parametrize(
-    "specification",
-    [
-        "simple_specification",
-        "simple_specification_with_pip",
-    ],
-)
-@mock.patch.object(generate_lockfile, "yaml", wraps=yaml)
-@mock.patch("conda_store_server._internal.action.generate_lockfile.logged_command")
-@mock.patch("conda_store_server._internal.action.generate_lockfile.run_lock")
-@pytest.mark.long_running_test
-def test_solve_lockfile(
-    mock_run_lock,
-    mock_logged_command,
-    mock_yaml,
-    conda_store,
-    specification,
-    request,
-):
-    """Test that the call to conda_lock.run_lock is formed correctly.
-
-    Mock out logged_command, which is used to call `conda info`; this is an
-    extremely slow call, and we aren't testing any aspect of it here.
-
-    Mock out the yaml package, it is used to load the lockfile normally
-    generated by conda-lock.
-    """
-
-    # Dump dummy data to the expected lockfile output location
-    def run_lock_side_effect(lockfile_path, **kwargs):
-        with open(lockfile_path, "w") as f:
-            yaml.dump({"foo": "bar"}, f)
-
-    mock_run_lock.side_effect = run_lock_side_effect
-
-    platforms = [conda_utils.conda_platform()]
-    specification = request.getfixturevalue(specification)
-    if specification.variables is None:
-        cuda_version = None
-    else:
-        cuda_version = specification.variables.get("CONDA_OVERRIDE_CUDA")
-
-    context = generate_lockfile.action_solve_lockfile(
-        conda_command=conda_store.conda_command,
-        specification=specification,
-        platforms=platforms,
-    )
-
-    # Check that the call to `conda_lock` is correctly formed
-    mock_run_lock.assert_called_once()
-    call_args = mock_run_lock.call_args_list[0][1]
-    assert str(call_args["environment_files"][0]).endswith("environment.yaml")
-    assert call_args["platforms"] == platforms
-    assert str(call_args["lockfile_path"]).endswith("conda-lock.yaml")
-    assert call_args["conda_exe"] == conda_store.conda_command
-    assert call_args["with_cuda"] == cuda_version
-
-    assert context.result["foo"] == "bar"
-
-
-def test_solve_lockfile_valid_conda_flags(conda_store, simple_specification):
-    context = action.action_solve_lockfile(
-        conda_command=conda_store.conda_command,
-        specification=simple_specification,
-        platforms=[conda_utils.conda_platform()],
-        conda_flags="--strict-channel-priority",
-    )
-    assert len(context.result["package"]) != 0
-
-
-# Checks that conda_flags is used by conda-lock
-@pytest.mark.long_running_test
-def test_solve_lockfile_invalid_conda_flags(conda_store, simple_specification):
-    with pytest.raises(
-        Exception, match=(r"Command.*--this-is-invalid.*returned non-zero exit status")
-    ):
-        action.action_solve_lockfile(
-            conda_command=conda_store.conda_command,
-            specification=simple_specification,
-            platforms=[conda_utils.conda_platform()],
-            conda_flags="--this-is-invalid",
-        )
-
-
-@pytest.mark.parametrize(
-    "specification",
-    [
-        "simple_specification",
-        "simple_specification_with_pip",
-    ],
-)
-@pytest.mark.long_running_test
-def test_solve_lockfile_multiple_platforms(conda_store, specification, request):
-    specification = request.getfixturevalue(specification)
-    context = action.action_solve_lockfile(
-        conda_command=conda_store.conda_command,
-        specification=specification,
-        platforms=["osx-64", "linux-64", "win-64", "osx-arm64"],
-    )
-    assert len(context.result["package"]) != 0
-
-
-def test_save_lockfile(simple_lockfile_specification):
-    """Ensure lockfile is saved in conda-lock `output` format"""
-    context = action.action_save_lockfile(specification=simple_lockfile_specification)
-    assert context.result == simple_lockfile_specification.lockfile.dict_for_output()
 
 
 @pytest.mark.parametrize(
