+++ conda-store-server/tests/plugins/lock/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) conda-store development team. All rights reserved.
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file.
